% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbase_fn_dom.R
\name{dbase_fn}
\alias{dbase_fn}
\title{Calculation of the number of degrees to basal species}
\usage{
dbase_fn(L)
}
\arguments{
\item{L}{}
}
\value{
A vector of the distance to the base for every species. Producers have a default distance of 1. Unconnected species have a distance of 0
}
\description{
Calculation of the number of degrees to basal species
}
\author{
Dominique Gravel
}
