% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe_fn.R
\name{lognormal_median}
\alias{lognormal_median}
\title{calculate a lognormal median}
\usage{
lognormal_median(n)
}
\arguments{
\item{n}{number of samples}
}
\value{

}
\description{
calculate a lognormal median
}
