% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_addition.R
\name{matrix_sum}
\alias{matrix_sum}
\title{Calculate the sum of elements in a numeric matrix}
\usage{
matrix_sum(x)
}
\arguments{
\item{x}{the matrix to be summed}
}
\value{
Integer
}
\description{
Calculate the sum of elements in a numeric matrix
}
