% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedro_data.R
\docType{data}
\name{pedro_data}
\alias{pedro_data}
\title{Pedro's dataset}
\format{
An object of class \code{numeric} of length 200.
}
\usage{
pedro_data
}
\description{
its just random numbers
}
\keyword{datasets}
