% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_fn_dom.R
\name{stats_fn}
\alias{stats_fn}
\title{Calculating statistics of food webs}
\usage{
stats_fn(x)
}
\arguments{
\item{x}{binary link matrix}
}
\value{
Something
}
\description{
Calculating statistics of food webs
}
