% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_addition.R
\name{sum_matrix}
\alias{sum_matrix}
\title{Sum a matrix}
\usage{
sum_matrix(matrix)
}
\arguments{
\item{matrix}{The matrix to be summed}
}
\value{
Integer
}
\description{
This function sums up one sumeric matrix
}
